fix: `EpochInfo` type import issue
