+++ packages/contracts-sdk/src/lib/types.ts
@@ -9,19 +9,3 @@ export type ValidatorStruct = {
   seconderPubkey: BigNumber;
   receiverPubkey: BigNumber;
 };
-
-/**
- * from the `getActiveUnkickedValidatorStructsAndCounts` Staking contract function
-   epochLength: _BigNumber { _hex: '0x05dc', _isBigNumber: true },
-  number: _BigNumber { _hex: '0x04c5', _isBigNumber: true },
-  endTime: _BigNumber { _hex: '0x66c75b12', _isBigNumber: true },
-  retries: _BigNumber { _hex: '0x03', _isBigNumber: true },
-  timeout: _BigNumber { _hex: '0x3c', _isBigNumber: true }
- */
-export type EpochInfo = {
-  epochLength: number;
-  number: number;
-  endTime: number;
-  retries: number;
-  timeout: number;
-};
