+++ packages/types/src/lib/types.ts
@@ -247,3 +247,19 @@ export interface LitContractResolverContext {
 }
 
 export type ResponseStrategy = 'leastCommon' | 'mostCommon' | 'custom';
+
+/**
+ * from the `getActiveUnkickedValidatorStructsAndCounts` Staking contract function
+   epochLength: _BigNumber { _hex: '0x05dc', _isBigNumber: true },
+  number: _BigNumber { _hex: '0x04c5', _isBigNumber: true },
+  endTime: _BigNumber { _hex: '0x66c75b12', _isBigNumber: true },
+  retries: _BigNumber { _hex: '0x03', _isBigNumber: true },
+  timeout: _BigNumber { _hex: '0x3c', _isBigNumber: true }
+ */
+export type EpochInfo = {
+  epochLength: number;
+  number: number;
+  endTime: number;
+  retries: number;
+  timeout: number;
+};
